requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.7.0
pandas>=1.0.0
numpy>=1.0.0
google-generativeai>=0.3.0 